checkpoint
dep reordering
