+++ bin/programs/client/Cargo.toml
@@ -13,8 +13,13 @@ cfg-if.workspace = true
 alloy-primitives.workspace = true
 alloy-consensus.workspace = true
 alloy-rlp.workspace = true
+alloy-eips.workspace = true
+op-alloy-consensus.workspace = true
 anyhow.workspace = true
 revm = { workspace = true, features = ["optimism"] }
+lru.workspace = true
+spin.workspace = true
+async-trait.workspace = true
 
 # local
 kona-common = { path = "../../../crates/common", version = "0.0.1" }
@@ -23,19 +28,6 @@ kona-preimage = { path = "../../../crates/preimage", version = "0.0.1" }
 kona-mpt = { path = "../../../crates/mpt", version = "0.0.1" }
 kona-derive = { path = "../../../crates/derive", version = "0.0.1" }
 
-# external
-op-alloy-consensus = { git = "https://github.com/clabby/op-alloy", branch = "refcell/consensus-port", default-features = false }
-alloy-eips = { git = "https://github.com/alloy-rs/alloy", rev = "cb95183", default-features = false }
-lru = "0.12.3"
-async-trait = "0.1.80"
-spin = "0.9.8"
-
 [dev-dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1.36.0", features = ["full"] }
-nybbles = "0.2.1"
-alloy-provider = { git = "https://github.com/alloy-rs/alloy", rev = "cb95183" }
-alloy-rpc-types = { git = "https://github.com/alloy-rs/alloy", rev = "cb95183" }
-alloy-transport-http = { git = "https://github.com/alloy-rs/alloy", rev = "cb95183" }
-reqwest = "0.12"
+serde = { version = "1.0.197", features = ["derive"] }
+serde_json = "1.0.117"
