+++ crates/mpt/Cargo.toml
@@ -19,10 +19,6 @@ revm.workspace = true
 
 # External
 alloy-trie = { version = "0.4.1", default-features = false }
-smallvec = "1.13"
-
-serde = "1.0"
-serde_json = "1.0"
 
 [dev-dependencies]
 alloy-consensus.workspace = true
