+++ crates/mpt/src/util.rs
@@ -1,7 +1,7 @@
 //! Utilities for `kona-mpt`
 
 use alloc::vec::Vec;
-use alloy_rlp::{BufMut, Encodable};
+use alloy_rlp::{Buf, BufMut, Encodable, Header};
 use alloy_trie::{proof::ProofRetainer, HashBuilder, Nibbles};
 
 /// Compute a trie root of the collection of items with a custom encoder.
@@ -48,4 +48,43 @@ pub(crate) const fn adjust_index_for_rlp(i: usize, len: usize) -> usize {
     } else {
         i + 1
     }
+}
+
+/// Walks through a RLP list's elements and returns the total number of elements in the list.
+/// Returns [alloy_rlp::Error::UnexpectedString] if the RLP stream is not a list.
+///
+/// ## Takes
+/// - `buf` - The RLP stream to walk through
+///
+/// ## Returns
+/// - `Ok(usize)` - The total number of elements in the list
+/// - `Err(_)` - The RLP stream is not a list
+pub(crate) fn rlp_list_element_length(buf: &mut &[u8]) -> alloy_rlp::Result<usize> {
+    let header = Header::decode(buf)?;
+    if !header.list {
+        return Err(alloy_rlp::Error::UnexpectedString);
+    }
+    let len_after_consume = buf.len() - header.payload_length;
+
+    let mut list_element_length = 0;
+    while buf.len() > len_after_consume {
+        let header = Header::decode(buf)?;
+        buf.advance(header.payload_length);
+        list_element_length += 1;
+    }
+    Ok(list_element_length)
+}
+
+/// Unpack node path to nibbles.
+///
+/// ## Takes
+/// - `first` - first nibble of the path if it is odd. Must be <= 0x0F, or will create invalid
+///   nibbles.
+/// - `rest` - rest of the nibbles packed
+///
+/// ## Returns
+/// - `Nibbles` - unpacked nibbles
+pub(crate) fn unpack_path_to_nibbles(first: Option<u8>, rest: &[u8]) -> Nibbles {
+    let rest = Nibbles::unpack(rest);
+    Nibbles::from_vec_unchecked(first.into_iter().chain(rest.iter().copied()).collect::<Vec<u8>>())
 }
