+++ crates/plasma/Cargo.toml
@@ -12,15 +12,13 @@ homepage.workspace = true
 # Workspace
 anyhow.workspace = true
 tracing.workspace = true
+alloy-consensus.workspace = true
+alloy-primitives = { workspace = true, features = ["rlp"] }
+async-trait.workspace = true
 
 # Local
 kona-primitives = { path = "../primitives" }
 kona-derive = { path = "../derive" }
-
-# External
-alloy-consensus = { git = "https://github.com/alloy-rs/alloy", rev = "cb95183", default-features = false }
-alloy-primitives = { workspace = true, features = ["rlp"] }
-async-trait = "0.1.80"
 
 # `serde` feature dependencies
 serde = { version = "1.0.197", default-features = false, features = ["derive"], optional = true }
