+++ crates/preimage/Cargo.toml
@@ -14,12 +14,10 @@ anyhow.workspace = true
 cfg-if.workspace = true
 tracing.workspace = true
 alloy-primitives.workspace = true
+async-trait.workspace = true
 
 # local
 kona-common = { path = "../common", version = "0.0.1" }
-
-# external
-async-trait = "0.1.80"
 
 [dev-dependencies]
 tokio = { version = "1.36.0", features = ["full"] }
